+++ test/problem_examples/digital_demo/testcases/1.sol
@@ -1,0 +1,1 @@
+demo_and_mux: passed